1.4.9 release
- Fix for PYFB-52
- Fix tests for PYFB-51 and PYFB-52

+++ fdb/fbcore.py
@@ -1254,7 +1254,7 @@ class Connection(object):
             uShortSize = struct.calcsize('<H')
             intSize = struct.calcsize('<i')
             pairSize = uShortSize + intSize
-            pairCount = len(buf) / pairSize
+            pairCount = int(len(buf) / pairSize)
 
             counts = {}
             for i in range(pairCount):
